Start porting to Go 1
Move the files around so that 'go build' and 'go test' and friends
work.  Write tests for the OID and Chunk code after updating them for
Go 1.

The types have been changed a little as well, using a pointer to a
fixed array for the OID type rather than a slice, should save two int
fields.