Linux-specific dir reading
In its attempt to be portable, Go's os.Readdir returns the entries
unsorted, and performs a stat.  On large directories, this can result
in a lot of seeking.

Implement a Readdir, more Posix specialized that returns just the name
and inode number returned by the readdir_r() library call.  This can
then be sorted by inode number before calling Lstat on the entries,
for example.