Implement ctime cache
The ctime cache stores all of the files (in the directory they are
present in) along with the ctime and data hash.  This allows for files
that don't change to not have to be reread.

The cache also contains an expiration time, which is set randomly
between 2-6 weeks.  Once the time expires, the file will be reread.
This helps spread out the rereading, but does eventually catch
problems.

This implements a 'cache regen' command which can generate the cache
from a previously made back.  It doesn't use the cache yet in the
backups themselves.